backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.2
PyMySQL>=1.1
pytest>=8.0
pytest-django>=4.8
pytest-html>=4.1
allure-pytest>=2.13
requests>=2.32
django-filter>=24.0
django-cors-headers>=4.4
drf-yasg>=1.21
python-dotenv>=1.0
